Refactoring of data workflows
