+++ .github/workflows/build.yml
@@ -19,6 +19,14 @@ jobs:
     steps:
     - uses: actions/checkout@v4
 
+    - name: Fetch up-to-date currency rates and icons
+      run: |
+        git fetch origin exchange-rates
+        git restore --source=FETCH_HEAD --worktree \
+          core/data/src/main/assets/crypto-rates.json \
+          core/data/src/main/assets/fiat-rates.json \
+          cryptoicons/src/main/res/drawable
+
     - name: Set up JDK 21
       uses: actions/setup-java@v4
       with:
@@ -54,6 +62,14 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
+
+    - name: Fetch up-to-date currency rates and icons
+      run: |
+        git fetch origin exchange-rates
+        git restore --source=FETCH_HEAD --worktree \
+          core/data/src/main/assets/crypto-rates.json \
+          core/data/src/main/assets/fiat-rates.json \
+          cryptoicons/src/main/res/drawable
 
     - name: Set up JDK 21
       uses: actions/setup-java@v4
@@ -100,6 +116,14 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - name: Fetch up-to-date currency rates and icons
+        run: |
+          git fetch origin exchange-rates
+          git restore --source=FETCH_HEAD --worktree \
+            core/data/src/main/assets/crypto-rates.json \
+            core/data/src/main/assets/fiat-rates.json \
+            cryptoicons/src/main/res/drawable
+
       - name: Set up JDK 21
         uses: actions/setup-java@v4
         with:
@@ -113,4 +137,3 @@ jobs:
         with:
           name: lint-results
           path: ./app/build/reports/*.html
-
