--- .github/workflows/daily-update-rates.yml
@@ -1,52 +1,0 @@
-name: Daily Update Rates
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "0 0,8,16 * * *"
-
-env:
-  OPENEXCHANGERATES_API_KEY: ${{ secrets.OPENEXCHANGERATES_API_KEY }}
-
-jobs:
-  update-rates:
-    runs-on: ubuntu-latest
-  
-    permissions:
-      contents: write
-
-    steps:
-      - name: Check out the repository
-        uses: actions/checkout@v4
-        with:
-          ref: exchange-rates
-      
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.x'
-
-      - name: Install dependencies
-        run: |
-          pip install --upgrade pip
-      
-      - name: Run Android rates update
-        run: python update_rates_and_icons.py --platform android
-          
-      - name: Check if OPENEXCHANGERATES_API_KEY secret is set
-        run: |
-          if [ -z "${{ secrets.OPENEXCHANGERATES_API_KEY }}" ]; then
-            echo "OPENEXCHANGERATES_API_KEY is NOT set"
-          else
-            echo "OPENEXCHANGERATES_API_KEY is set (value is masked)"
-          fi
-       
-      - name: Push into Git
-        run: |-
-          git config --global user.email "bot@ark-builders.dev"
-          git config --global user.name "ARK Builders Bot"
-          git add core/data/src/main/assets/fiat-rates.json
-          git add core/data/src/main/assets/crypto-rates.json
-          git commit -m "Automatic update of rates" || echo "Nothing to commit"
-          git pull origin exchange-rates --no-edit
-          git push origin exchange-rates
